Move shared quick fix implementations to IssueQuickFix.h
