+++ common/src/Model/IssueQuickFix.cpp
@@ -20,6 +20,8 @@
 #include "IssueQuickFix.h"
 
 #include "Model/Issue.h"
+#include "Model/MapFacade.h"
+#include "Model/PushSelection.h"
 
 #include <cassert>
 #include <string>
@@ -51,6 +53,54 @@ void IssueQuickFix::doApply(MapFacade* facade, const std::vector<const Issue*>& 
 
 void IssueQuickFix::doApply(MapFacade*, const Issue&) const {
   assert(false);
+}
+
+RemoveEntityPropertiesQuickFix::RemoveEntityPropertiesQuickFix(const IssueType issueType)
+  : IssueQuickFix(issueType, "Delete properties") {}
+
+void RemoveEntityPropertiesQuickFix::doApply(MapFacade* facade, const Issue& issue) const {
+  const PushSelection push(facade);
+
+  const auto& propIssue = static_cast<const EntityPropertyIssue&>(issue);
+
+  // If world node is affected, the selection will fail, but if nothing is selected,
+  // the removeProperty call will correctly affect worldspawn either way.
+
+  facade->deselectAll();
+  facade->selectNodes({&issue.node()});
+  facade->removeProperty(propIssue.propertyKey());
+}
+
+TransformEntityPropertiesQuickFix::TransformEntityPropertiesQuickFix(
+  const IssueType issueType, const std::string& description, const KeyTransform& keyTransform,
+  const ValueTransform& valueTransform)
+  : IssueQuickFix(issueType, description)
+  , m_keyTransform(keyTransform)
+  , m_valueTransform(valueTransform) {}
+
+void TransformEntityPropertiesQuickFix::doApply(MapFacade* facade, const Issue& issue) const {
+  const PushSelection push(facade);
+
+  const auto& propIssue = static_cast<const EntityPropertyIssue&>(issue);
+  const auto& oldkey = propIssue.propertyKey();
+  const auto& oldValue = propIssue.propertyValue();
+  const auto newKey = m_keyTransform(oldkey);
+  const auto newValue = m_valueTransform(oldValue);
+
+  // If world node is affected, the selection will fail, but if nothing is selected,
+  // the removeProperty call will correctly affect worldspawn either way.
+
+  facade->deselectAll();
+  facade->selectNodes({&issue.node()});
+
+  if (newKey.empty()) {
+    facade->removeProperty(propIssue.propertyKey());
+  } else {
+    if (newKey != oldkey)
+      facade->renameProperty(oldkey, newKey);
+    if (newValue != oldValue)
+      facade->setProperty(newKey, newValue);
+  }
 }
 } // namespace Model
 } // namespace TrenchBroom
