+++ common/src/Model/LongPropertyKeyValidator.cpp
@@ -23,8 +23,8 @@
 #include "Model/Entity.h"
 #include "Model/EntityNode.h"
 #include "Model/Issue.h"
+#include "Model/IssueQuickFix.h"
 #include "Model/MapFacade.h"
-#include "Model/RemoveEntityPropertiesQuickFix.h"
 
 #include <string>
 #include <vector>
